refactor(utils): split 'utils' by functionality
