+++ src/mcp_server_uyuni/utils/__init__.py
@@ -1,0 +1,12 @@
+from .elicitation import client_supports_elicitation, elicit_approval
+from .pagination import normalize_pagination, build_list_meta, paginate_items
+from .misc import matches_optional_filter
+
+__all__ = [
+    "client_supports_elicitation",
+    "elicit_approval",
+    "normalize_pagination",
+    "build_list_meta",
+    "paginate_items",
+    "matches_optional_filter",
+]
